# swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Henohenoapi
  description: >-
    The project aims to create a platform for managing exams, teachers can
    public exams(which have a deadline) and add tasks to them, while students
    can submit their answers and answer.
#dominio e baseroot
host: henoheno-app.herokuapp.com
basePath: /
...
schemes:
  - http
#formati dati accettabile da server
consumes:
  - application/json
#formati dati prodotto da server
produces:
  - application/json
...
security:
  - Api_key_student: []     #apikey di studente
  - Api_key_teacher: []     #apikey di insegnante



#parte di url--------------------------------------------------------------------------
paths:
...
  '/users/{user_id}/exams':
    get:
      summary: get the exams assigned to a student
      description: get the exams assigned to a student
      tags:
        - Users_resource
      
      parameters:
        - $ref: '#/parameters/User_id_in_path'
        - in: query
          name: selection
          schema:
            type: string
          description: Assumes values "created" or "assigned"
      responses:
                '200':
                        description: OK
                        schema:
                              $ref: '#/definitions/Exams'  
                '401':
                        $ref: '#/responses/401_unauthorized'
                '404':
                        $ref: '#/responses/404_not_found'
                '500':
                        $ref: '#/responses/500_internal_server_error'
  '/users/{user_id}/exam_submissions':
    get:
      summary: get the submissions of a student
      description: get the submissions of a student
      tags:
        - Users_resource
      security:
        - Api_key_teacher: []  #solo insegnante
      parameters:
        - $ref: '#/parameters/User_id_in_path'
      responses:
                '200':
                        description: OK
                        schema:
                              $ref: '#/definitions/Exam_submissions' 
                '401':
                        $ref: '#/responses/401_unauthorized'
                '404':
                        $ref: '#/responses/404_not_found'
                '500':
                        $ref: '#/responses/500_internal_server_error'
...
  /exams:
    get:
      summary: exams list
      description: 
        Teachers get the list of created exams, while students the list of assigned exams(the tasks solutions will be empty in this case)
      tags:
        - Exams_resource
      responses:
        '200':
          description: OK
          schema:
              $ref: '#/definitions/Exams'
          
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'

    post:
      summary: create an exam by defining a set of tasks and other stuff
      description: create an exam by defining a set of tasks and other stuff
      tags:
        - Exams_resource
      security:
        - Api_key_teacher: []  #solo insegnante 
      parameters:
        - $ref: '#/parameters/Exam_in_post'
      responses:
        '201':
                description: Created
                schema:
                        type: integer
                        description: the new exam's id
        '204':
                $ref: '#/responses/204_no_content'
        '400':
                $ref: '#/responses/400_bad_request'
        '401':
                $ref: '#/responses/401_unauthorized'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/exams/{id}':
    get:
      summary: access one specific exam
      description: access one specific exam and its tasks
      tags:
        - Exams_resource
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam'
          
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'

    put:
      summary: update an exam by adding groups or changing specific fields
      description: update an exam by adding groups or changing specific fields
      tags:
        - Exams_resource
      security:
        - Api_key_teacher: []  #solo insegnante
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
        - $ref: '#/parameters/Exam_in_put'
      responses:
        '200':
                $ref: '#/responses/200_ok_only'
        '201':
                $ref: '#/responses/201_created'
        '204':
                $ref: '#/responses/204_no_content'
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '409':
                $ref: '#/responses/409_conflict'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/exams/{id}/exam_submissions':
    get:
      summary: access the answer of a specific exam
      description: get a exam submission list of a specific exam
      tags:
        - Exams_resource
      security:
        - Api_key_teacher: []  #solo insegnante
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
      responses:
        '200':
                description: OK
                schema:
                        $ref: '#/definitions/Exam_submissions' 
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/exams/{exam_id}/exam_peer_reviews':
    get:
      summary: access the reviews of a specific exam
      description: get the list of reviews of an exam
      tags:
        - Exams_resource
      security:
        - Api_key_teacher: []  #solo insegnante
      parameters:
        - $ref: '#/parameters/Exam_id_in_path'
      responses:
        '200':
                description: OK
                schema:
                        $ref: '#/definitions/Exam_peer_reviews' 
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
...
  /exam_submissions:
    get:
      summary: Exam submission list
      description: get the list of the submissions of your students or, if you're a student, get the list of all your submissions(tasks won't contain solutions)
      tags:
        - Exam_submissions_resource
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_submissions'
          
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
        
          
    post:
      summary: submit new answer
      description: submit new answer
      tags:
        - Exam_submissions_resource
      security:
         - Api_key_student: []     #solo lo studente
      parameters:
        - $ref: '#/parameters/Exam_submission_in_post'
         
      responses:
        '201':
                description: Created
                schema:
                        type: integer
                        description: the new exam submission's id
        '204':
                $ref: '#/responses/204_no_content'
        '400':
                $ref: '#/responses/400_bad_request'
        '401':
                $ref: '#/responses/401_unauthorized'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/exam_submissions/{exam_submission_id}':  
    get:  
      summary: access a specific answer
      description: read the submission of a student and its evaluation(if present)
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_submission'
          
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'
        
          
    put:
      summary: update an existing solution(or assign evaluation)
      description: As a student you might update your submission, or if you're a teacher you can set an evaluation for the submission
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
        - $ref: '#/parameters/Exam_submission_in_put'
      responses:
        '200':
                $ref: '#/responses/200_ok_only'
        '201':
                $ref: '#/responses/201_created'
        '204':
                $ref: '#/responses/204_no_content'
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '409':
                $ref: '#/responses/409_conflict'
        '500':
                $ref: '#/responses/500_internal_server_error'
  '/exam_submissions/{exam_submission_id}/exam_peer_reviews':
    get:
      summary: access the reviews of a solution
      description: get all reviews of a specific submission(if it's a student he won't get the solutions)
      tags:
        - Exam_submissions_resource
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Exam_peer_reviews'
          
        '401':
                $ref: '#/responses/401_unauthorized'
        '404':
                $ref: '#/responses/404_not_found'
        '500':
                $ref: '#/responses/500_internal_server_error'

        
          
    post:
      summary: submit a review to an answer
      description: submit a review to an answer
      tags:
        - Exam_submissions_resource
      security:
         - Api_key_student: []     #solo lo studente
      parameters:
        - $ref: '#/parameters/Exam_submission_id_in_path'
        - $ref: '#/parameters/Exam_peer_review_in_post'
      responses:
        '201':
                description: Created
                schema:
                        type: integer
                        description: the new exam peer review's id
        '204':
                $ref: '#/responses/204_no_content'
        '400':
                $ref: '#/responses/400_bad_request'
        '401':
                $ref: '#/responses/401_unauthorized'
        '500':
                $ref: '#/responses/500_internal_server_error'
...
definitions:
  Error:
    properties:
      code:
        type: integer
      message:
        type: string
  
###oggetto user
...
  Exam:
    type: object
    properties:
      id:  #primary key
        type: integer 
      owner: #foreign key
        type: integer
      title:
        type: string
      description:
        type: string
      taskset:
        type: array
        items:
          $ref: '#/definitions/Task_in_exam'
      group:
        $ref: '#/definitions/Group'
      final_deadline:
        type: string
        format: date
      review_deadline:
        type: string
        format: date
  Exams:
        type: array
        items: 
                $ref: '#/definitions/Exam'
  
###oggetto task
...
  Task_in_exam:
         type: object
         properties:
                task_id: #primary key
                       type: integer
                description:
                       type: string
         
                
###oggetto group
  Group:
    type: object
    properties:
      group_id: #primary key
        type: integer
      owner_id: #foreign key
        type: integer
      name:
        type: string
      description:
        type: string
      members:
        type: array
        items:
          $ref: '#/definitions/User_in_response'
...
  Exam_submission:
    type: object
    properties:
      exam_submission_id: #primary key
        type: integer
      referred_exam: #'foreign' key
        $ref: '#/definitions/Exam'
      submitter_id: #foreign key
        type: integer
      answers:
        type: array
        items:
          type: object
          properties:
            task_id:
              type: integer
            answer:
              type: string
      status:
        type: string
      evaluation:
        type: string
  Exam_submissions:
          type: array
          items:
                  $ref: '#/definitions/Exam_submission'  
###oggetto exam_peer_review
  Exam_peer_review:
    type: object
    properties:
      exam_peer_review_id: #primary key
        type: integer
      reviewer_id: #foreign key
        type: integer 
      submitted_exam: #'foreign' key
        $ref: '#/definitions/Exam_submission'
      review:
        type: array
        items:
          type: string
  Exam_peer_reviews:
          type: array
          items:
                  $ref: '#/definitions/Exam_peer_review'  
###additional objects##
  User_in_response:
      type: object   
      properties:
        user_id: 
          type: integer
          description: the user 's id
        email:
          type: string
          description: the user 's email
        name:
          type: string
          description: the user 's name
        surname:
          type: string
          description: the user 's surname
...
  Exam_for_post:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
      tasks_ids:
        type: array
        items:
            type: integer
      group_id:
        type: integer
      final_deadline:
        type: string
        format: date
      review_deadline:
        type: string
        format: date
...
  Exam_submission_for_post:
    type: object
    properties:
      referred_exam_id: #'foreign' key
        type: integer
      answers:
        type: array
        items:
          type: object
          properties:
            task_id:
              type: integer
            answer:
              type: string
      status:
        type: string
  Exam_submission_for_put:
    type: object
    properties:
      answers:
        type: array
        items:
          type: object
          properties:
            task_id:
              type: integer
            answer:
              type: string
      status:
        type: string
      evaluation:
        type: string
  Exam_peer_review_for_post:
    type: object
    properties:
      review:
        type: array
        items:
          type: string
...
parameters:
...
  User_id_in_path:
    name: user_id
    description: the user's id
    in: path
    type: integer
    required: true
  Exam_in_post:        
        name: exam
        description: the object exam
        in: body
        schema:
                $ref: '#/definitions/Exam_for_post'
                required:
                        - title
                        - description
                        - taskset
                        - final_deadline
                        - review_deadline
  Exam_in_put:        
        name: exam
        description: the object exam
        in: body
        schema:
                $ref: '#/definitions/Exam_for_post'
  Exam_id_in_path:
    name: id
    description: id esame
    in: path
    type: integer
    required: true
...
  Exam_submission_in_post:
        name: exam_submission
        description: the object exam_submission
        in: body
        schema:
                $ref: '#/definitions/Exam_submission_for_post'
                required:
                        - referred_exam_id
                        #- submitter
                        - answers
                        - status    
  Exam_submission_in_put:
        name: exam_submission
        description: the object exam_submission
        in: body
        schema:
                $ref: '#/definitions/Exam_submission_for_put'
  Exam_submission_id_in_path:
    name: exam_submission_id
    description: exam submission id
    in: path
    type: integer
    required: true  
  Exam_peer_review_in_post:
        name: exam_peer_review
        description: the object exam peer review
        in: body
        schema:
                $ref: '#/definitions/Exam_peer_review_for_post'
                required:
                        - review  
...
responses:
...
  200_ok_only:
        description: Created  
  201_created:
        description: Created
  204_no_content:
        description: No Content
  400_bad_request:
          description: Bad Request
          schema:
                $ref: '#/definitions/Error'
  401_unauthorized:
          description: Unauthorized, missing or invalid API Key
          schema:
                $ref: '#/definitions/Error'
  404_not_found:
          description: Not Found
          schema:
                 $ref: '#/definitions/Error'
...
  409_conflict:
          description: Conflict
          schema:
                 $ref: '#/definitions/Error'
  500_internal_server_error:
          description: Internal Server Error
          schema:
                $ref: '#/definitions/Error'
         
         
#definisce apikey      
securityDefinitions:
        Api_key_student:
                type: apiKey
                in: query
                name: token
        Api_key_teacher:
                type: apiKey
                in: query
                name: tokent